Dragging back to the same block will now not show the input

+++ js/block.js
@@ -1134,6 +1134,9 @@ function endDragBlock(evt){
         // No legal target, probably outside of workspace
         return cancelDragBlock();
     }
+    if(dropTarget === originalParent){
+        return cancelDragBlock();
+    }
     if (dropTarget === BLOCK_MENU){
         // Drop on script menu to delete block, always delete clone
         deleteOriginalBlock(originalBlock, originalParent, nextElem);
@@ -1201,9 +1204,12 @@ function resetDragging(){
     }
     if (origTarget){
         origTarget.classList.remove('hide');
-        var siblingInput = dom.child(origTarget.parentElement, 'input, select');
-        if (siblingInput){
-            siblingInput.classList.add('hide');
+        if(origTarget.parentElement){
+            // Hide the sibling input
+            var siblingInput = dom.child(origTarget.parentElement, 'input, select');
+            if (siblingInput){
+                siblingInput.classList.add('hide');
+            }
         }
     }
     dragTarget = null;
